Rename CacheSupervisor to Supervisor
