+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- BREAKING: Rename `ShopifyAPI.CacheSupervisor` to `ShopifyAPI.Supervisor`.
+
 ## 0.9.4
 
 - "/install" route now handles app name in path - thanks @tres
